+++ .travis.yml
@@ -25,8 +25,10 @@ install:
 script:
   # Test with the defaults
   - sbt new file://$(pwd) --name=hello && pushd hello && sbt test && popd
-    # Test with custom name/organization to ensure templating is done correctly
-  - sbt new file://$(pwd) --name=smello --organization=org.eggsample && cd smello && sbt test
+  # Test with custom name/organization to ensure templating is done correctly
+  - sbt new file://$(pwd) --name=smello --organization=org.eggsample && pushd smello && sbt test && popd
+  # Test Maven Archetype
+  - ./test-archetype.sh
 
 cache:
   directories:
