+++ src/main/g8/$if(!mavenArchetype.truthy)$.sbtopts$endif$
