+++ src/main/g8/$if(!mavenArchetype.truthy)$README.md$endif$
