+++ src/main/g8/$if(!mavenArchetype.truthy)$build.sbt$endif$
