+++ src/main/g8/$if(!mavenArchetype.truthy)$project$endif$/$if(!mavenArchetype.truthy)$plugins.sbt$endif$
