+++ src/main/g8/$if(mavenArchetype.truthy)$archetype-resources$endif$/$if(mavenArchetype.truthy)$__serviceName__$else$$name__norm$$endif$-api/src/main/java/$if(mavenArchetype.truthy)$__packageInPathFormat__$else$$package$$endif$/$if(mavenArchetype.truthy)$__serviceName__$endif$/api/GreetingMessage.java
@@ -1,0 +1,17 @@
+package $if(mavenArchetype.truthy)$\${package}.\${serviceName}$else$$package$$endif$.api;
+
+import com.fasterxml.jackson.annotation.JsonCreator;
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
+import com.google.common.base.Preconditions;
+import lombok.Value;
+
+@Value
+@JsonDeserialize
+public final class GreetingMessage {
+    public final String message;
+
+    @JsonCreator
+    public GreetingMessage(String message) {
+        this.message = Preconditions.checkNotNull(message, "message");
+    }
+}
