+++ src/main/g8/$if(mavenArchetype.truthy)$archetype-resources$endif$/$if(mavenArchetype.truthy)$__serviceName__$else$$name__norm$$endif$-impl/src/main/java/$if(mavenArchetype.truthy)$__packageInPathFormat__$else$$package$$endif$/$if(mavenArchetype.truthy)$__serviceName__$endif$/impl/$if(mavenArchetype.truthy)$__serviceClassName__$else$$name__Camel$$endif$State.java
@@ -1,0 +1,24 @@
+package $if(mavenArchetype.truthy)$\${package}.\${serviceName}$else$$package$$endif$.impl;
+
+import com.fasterxml.jackson.annotation.JsonCreator;
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
+import com.google.common.base.Preconditions;
+import com.lightbend.lagom.serialization.CompressedJsonable;
+import lombok.Value;
+
+/**
+ * The state for the {@link $if(mavenArchetype.truthy)$\${serviceClassName}$else$$name;format="Camel"$$endif$Entity} entity.
+ */
+@SuppressWarnings("serial")
+@Value
+@JsonDeserialize
+public final class $if(mavenArchetype.truthy)$\${serviceClassName}$else$$name;format="Camel"$$endif$State implements CompressedJsonable {
+    public final String message;
+    public final String timestamp;
+
+    @JsonCreator
+    $if(mavenArchetype.truthy)$\${serviceClassName}$else$$name;format="Camel"$$endif$State(String message, String timestamp) {
+        this.message = Preconditions.checkNotNull(message, "message");
+        this.timestamp = Preconditions.checkNotNull(timestamp, "timestamp");
+    }
+}
