+++ src/main/g8/$if(mavenArchetype.truthy)$archetype-resources$endif$/$if(mavenArchetype.truthy)$__serviceName__$else$$name__norm$$endif$-impl/src/main/resources/application.conf
@@ -1,0 +1,20 @@
+play.modules.enabled += $if(mavenArchetype.truthy)$\${package}.\${serviceName}$else$$package$$endif$.impl.$if(mavenArchetype.truthy)$\${serviceClassName}$else$$name;format="Camel"$$endif$Module
+
+$name;format="norm"$.cassandra.keyspace = $name;format="lower,snake,word"$
+
+$!
+The unusual syntax below is because this file is proccessed by StringTemplate.
+We are using StringTemplate properties to construct a config key name,
+which is then used in a HOCON substitution.
+
+The results, if name = hello, should look like this:
+
+cassandra-journal.keyspace = ${hello.cassandra.keyspace}
+cassandra-snapshot-store.keyspace = ${hello.cassandra.keyspace}
+lagom.persistence.read-side.cassandra.keyspace = ${hello.cassandra.keyspace}
+
+Also note that this comment is a StringTemplate comment and is not included in the output.
+!$
+cassandra-journal.keyspace = \${$name;format="norm"$.cassandra.keyspace}
+cassandra-snapshot-store.keyspace = \${$name;format="norm"$.cassandra.keyspace}
+lagom.persistence.read-side.cassandra.keyspace = \${$name;format="norm"$.cassandra.keyspace}
