+++ src/main/g8/$if(mavenArchetype.truthy)$archetype-resources$endif$/.gitignore
@@ -1,0 +1,15 @@
+target
+ideatarget
+.target
+bin
+.cache
+.cache-main
+.cache-tests
+.classpath
+.project
+.tmpBin
+.factorypath
+.settings
+logs
+.idea
+*.iml
