+++ src/main/g8/$if(mavenArchetype.truthy)$archetype-resources$endif$/LICENSE
