--- src/main/g8/$name__norm$-api/src/main/java/$package$/api/$name__Camel$Event.java
@@ -1,27 +1,0 @@
-package $package$.api;
-
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonSubTypes;
-import com.fasterxml.jackson.annotation.JsonTypeInfo;
-import com.google.common.base.Preconditions;
-import lombok.Value;
-
-@JsonTypeInfo(use = JsonTypeInfo.Id.NAME, property = "type")
-@JsonSubTypes({
-        @JsonSubTypes.Type(value = $name;format="Camel"$Event.GreetingMessageChanged.class, name = "greeting-message-changed")
-})
-public interface $name;format="Camel"$Event {
-    String getName();
-
-    @Value
-    final class GreetingMessageChanged implements $name;format="Camel"$Event {
-        public final String name;
-        public final String message;
-
-        @JsonCreator
-        public GreetingMessageChanged(String name, String message) {
-            this.name = Preconditions.checkNotNull(name, "name");
-            this.message = Preconditions.checkNotNull(message, "message");
-        }
-    }
-}
