--- src/main/g8/$name__norm$-api/src/main/java/$package$/api/GreetingMessage.java
@@ -1,17 +1,0 @@
-package $package$.api;
-
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-import com.google.common.base.Preconditions;
-import lombok.Value;
-
-@Value
-@JsonDeserialize
-public final class GreetingMessage {
-    public final String message;
-
-    @JsonCreator
-    public GreetingMessage(String message) {
-        this.message = Preconditions.checkNotNull(message, "message");
-    }
-}
