--- src/main/g8/$name__norm$-impl/src/test/java/$package$/impl/$name__Camel$ServiceTest.java
@@ -1,30 +1,0 @@
-package $package$.impl;
-
-import org.junit.Test;
-
-import $package$.api.GreetingMessage;
-import $package$.api.$name;format="Camel"$Service;
-
-import static com.lightbend.lagom.javadsl.testkit.ServiceTest.defaultSetup;
-import static com.lightbend.lagom.javadsl.testkit.ServiceTest.withServer;
-import static java.util.concurrent.TimeUnit.SECONDS;
-import static org.junit.Assert.assertEquals;
-
-public class $name;format="Camel"$ServiceTest {
-    @Test
-    public void shouldStorePersonalizedGreeting() {
-        withServer(defaultSetup().withCassandra(), server -> {
-            $name;format="Camel"$Service service = server.client($name;format="Camel"$Service.class);
-
-            String msg1 = service.hello("Alice").invoke().toCompletableFuture().get(5, SECONDS);
-            assertEquals("Hello, Alice!", msg1); // default greeting
-
-            service.useGreeting("Alice").invoke(new GreetingMessage("Hi")).toCompletableFuture().get(5, SECONDS);
-            String msg2 = service.hello("Alice").invoke().toCompletableFuture().get(5, SECONDS);
-            assertEquals("Hi, Alice!", msg2);
-
-            String msg3 = service.hello("Bob").invoke().toCompletableFuture().get(5, SECONDS);
-            assertEquals("Hello, Bob!", msg3); // default greeting
-        });
-    }
-}
