--- src/main/g8/$name__norm$-stream-impl/src/main/resources/application.conf
@@ -1,38 +1,0 @@
-play.modules.enabled += $package$stream.impl.$name;format="Camel"$StreamModule
-
-$name;format="norm"$-stream.cassandra.keyspace = $name;format="lower,snake,word"$_stream
-
-$!
-The unusual syntax below is because this file is proccessed by StringTemplate.
-We are using StringTemplate properties to construct a config key name,
-which is then used in a HOCON substitution.
-
-The results, if name = hello, should look like this:
-
-cassandra-journal.keyspace = ${hello-stream.cassandra.keyspace}
-cassandra-snapshot-store.keyspace = ${hello-stream.cassandra.keyspace}
-lagom.persistence.read-side.cassandra.keyspace = ${hello-stream.cassandra.keyspace}
-
-Also note that this comment is a StringTemplate comment and is not included in the output.
-!$
-cassandra-journal.keyspace = \${$name;format="norm"$-stream.cassandra.keyspace}
-cassandra-snapshot-store.keyspace = \${$name;format="norm"$-stream.cassandra.keyspace}
-lagom.persistence.read-side.cassandra.keyspace = \${$name;format="norm"$-stream.cassandra.keyspace}
-
-
-# The properties below override Lagom default configuration with the recommended values for new projects.
-#
-# Lagom has not yet made these settings the defaults for backward-compatibility reasons.
-
-# Prefer 'ddata' over 'persistence' to share cluster sharding state for new projects.
-# See https://doc.akka.io/docs/akka/current/cluster-sharding.html#distributed-data-vs-persistence-mode
-akka.cluster.sharding.state-store-mode = ddata
-
-# Enable the serializer provided in Akka 2.5.8+ for akka.Done and other internal
-# messages to avoid the use of Java serialization.
-akka.actor.serialization-bindings {
-  "akka.Done"                 = akka-misc
-  "akka.NotUsed"              = akka-misc
-  "akka.actor.Address"        = akka-misc
-  "akka.remote.UniqueAddress" = akka-misc
-}
