--- src/main/g8/.gitignore
@@ -1,13 +1,0 @@
-target
-ideatarget
-.target
-bin
-.cache
-.cache-main
-.cache-tests
-.classpath
-.project
-.tmpBin
-.factorypath
-.settings
-logs
