+++ src/main/g8/default.properties
@@ -2,3 +2,5 @@ name = Hello
 organization = com.example
 version = 1.0-SNAPSHOT
 package = $organization$.$name;format="lower,word"$
+mavenArchetype = no
+verbatim = archetype-metadata.xml *pom.xml*
